+++ cookbooks/chub_sensu/attributes/default.rb
@@ -19,3 +19,7 @@ default[:chub_sensu][:redis][:check_redis_root] = "/opt/check_redis_master"
 default[:chub_sensu][:redis][:check_redis_file] = "redis-role.sh"
 
 default[:sensu][:rabbitmq][:heartbeat] = 120
+default[:sensu][:init_style] = "runit"
+
+default[:redisio][:default_settings][:logfile] = "/var/log/redis.log"
+default[:redisio][:sentinel_defaults][:logfile] = "/var/log/redis_sentinel.log"