+++ cookbooks/chub_sensu/CHANGES.md
@@ -1,3 +1,11 @@
+# 0.1.20
+
+Larry - Moving redis log into log instead of syslog.
+
+# 0.1.19
+
+Larry - Addition of service upgrade recipe
+
 # 0.1.18
 
 Larry - Allow cookbook to use FQDN over ip address